+++ doc/usage/walkthrough.clj
@@ -1,5 +1,6 @@
-(ns usage.core-wrap-handler
-  (:require [strojure.ring-stack.core :as stack]))
+(ns usage.walkthrough
+  (:require [strojure.ring-stack.handler :as handler]
+            [strojure.ring-stack.middleware :as mid]))
 
 ;; ## Define middleware functions
 
@@ -21,7 +22,7 @@
               (update :trace/response conj 'wrap2))
       (->> (println 'wrap2)))))
 
-(stack/set-as-handler-fn `wrap2 (constantly wrap2))
+(mid/set-handler-fn `wrap2 (constantly wrap2))
 
 (defn- wrap3 [handler]
   (fn [request]
@@ -40,10 +41,10 @@
       (->> (println 'wrap4)))))
 
 ;; Register functions to be used in configuration.
-(stack/set-as-handler-fn `wrap1 (constantly wrap1) {:type-aliases [wrap1 ::wrap1]})
-(stack/set-as-handler-fn `wrap2 (constantly wrap2) {:type-aliases [wrap2 ::wrap2]})
-(stack/set-as-handler-fn `wrap3 (constantly wrap3) {:type-aliases [wrap3 ::wrap3]})
-(stack/set-as-handler-fn `wrap4 (constantly wrap4) {:type-aliases [wrap4 ::wrap4]})
+(mid/set-handler-fn `wrap1 (constantly wrap1) {:type-aliases [wrap1 ::wrap1]})
+(mid/set-handler-fn `wrap2 (constantly wrap2) {:type-aliases [wrap2 ::wrap2]})
+(mid/set-handler-fn `wrap3 (constantly wrap3) {:type-aliases [wrap3 ::wrap3]})
+(mid/set-handler-fn `wrap4 (constantly wrap4) {:type-aliases [wrap4 ::wrap4]})
 
 ;; - Define ring request wrappers
 
@@ -56,8 +57,8 @@
   (update request :trace/request conj 'request2))
 
 ;; Register functions to be used in configuration.
-(stack/set-as-request-fn `request1 (constantly request1) {:type-aliases [request1 ::request1]})
-(stack/set-as-request-fn `request2 (constantly request2) {:type-aliases [request2 ::request2]})
+(mid/set-request-fn `request1 (constantly request1) {:type-aliases [request1 ::request1]})
+(mid/set-request-fn `request2 (constantly request2) {:type-aliases [request2 ::request2]})
 
 ;; - Define ring response wrappers
 
@@ -70,8 +71,8 @@
   (update response :trace/response conj 'response2))
 
 ;; Register functions to be used in configuration.
-(stack/set-as-response-fn `response1 (constantly response1) {:type-aliases [response1 ::response1]})
-(stack/set-as-response-fn `response2 (constantly response2) {:type-aliases [response2 ::response2]})
+(mid/set-response-fn `response1 (constantly response1) {:type-aliases [response1 ::response1]})
+(mid/set-response-fn `response2 (constantly response2) {:type-aliases [response2 ::response2]})
 
 ;; ## Define handler function to be wrapped
 
@@ -84,10 +85,10 @@
 
 ;; - Compose handler functions
 
-(let [handler (stack/wrap-handler handler* {:outer [wrap1 wrap2]
-                                            :enter [request1 request2]
-                                            :leave [response1 response2]
-                                            :inner [wrap3 wrap4]})]
+(let [handler (handler/build handler* {:outer [wrap1 wrap2]
+                                       :enter [request1 request2]
+                                       :leave [response1 response2]
+                                       :inner [wrap3 wrap4]})]
   (handler {:trace/request []}))
 
 ;wrap1 #:trace{:request []}
@@ -109,42 +110,42 @@
 
 ;; - Compose handler types (symbols)
 
-(let [handler (stack/wrap-handler handler* {:outer [`wrap1 `wrap2]
-                                            :enter [`request1 `request2]
-                                            :leave [`response1 `response2]
-                                            :inner [`wrap3 `wrap4]})]
+(let [handler (handler/build handler* {:outer [`wrap1 `wrap2]
+                                       :enter [`request1 `request2]
+                                       :leave [`response1 `response2]
+                                       :inner [`wrap3 `wrap4]})]
   (handler {:trace/request []}))
 
 ;; - Compose handler types (keywords)
 
-(let [handler (stack/wrap-handler handler* {:outer [::wrap1 ::wrap2]
-                                            :enter [::request1 ::request2]
-                                            :leave [::response1 ::response2]
-                                            :inner [::wrap3 ::wrap4]})]
+(let [handler (handler/build handler* {:outer [::wrap1 ::wrap2]
+                                       :enter [::request1 ::request2]
+                                       :leave [::response1 ::response2]
+                                       :inner [::wrap3 ::wrap4]})]
   (handler {:trace/request []}))
 
 ;; - Middleware with dependency
 
 ;; Requires to :enter `request1` before `request2`
-(stack/set-required-config `request2 {:enter [`request1]})
+(mid/set-required-config `request2 {:enter [`request1]})
 
 (comment
   ;; Missing required middleware
-  (let [handler (stack/wrap-handler handler* {:enter [request2]})]
+  (let [handler (mid/wrap-handler handler* {:enter [request2]})]
     (handler {:trace/request []}))
   ;clojure.lang.ExceptionInfo:
   ; Missing required middleware: {:middleware usage.core_wrap_handler$request2, :requires usage.core-wrap-handler/request1} {:middleware usage.core_wrap_handler$request2, :required-config {:enter [usage.core-wrap-handler/request1]}, :missing usage.core-wrap-handler/request1}
 
   ;; Required middleware in wrong position
-  (let [handler (stack/wrap-handler handler* {:enter [request2
-                                                      request1]})]
+  (let [handler (mid/wrap-handler handler* {:enter [request2
+                                                    request1]})]
     (handler {:trace/request []}))
   ;clojure.lang.ExceptionInfo:
   ; Required middleware in wrong position: {:middleware usage.core_wrap_handler$request2, :requires usage.core-wrap-handler/request1} {:middleware usage.core_wrap_handler$request2, :required-config {:enter [usage.core-wrap-handler/request1]}, :missing usage.core-wrap-handler/request1}
 
   ;; Ignore dependency error
-  (let [handler (stack/wrap-handler handler* {:enter [request2]
-                                              :ignore-required #{request1}})]
+  (let [handler (mid/wrap-handler handler* {:enter [request2]
+                                            :ignore-required #{request1}})]
     (handler {:trace/request []}))
   ;request2 #:trace{:request []}
   ;handler* #:trace{:request [request2]}